Move pipeline utils into its own package ('substation')
