+++ openknotscore/substation/README.md
@@ -1,0 +1,4 @@
+# Substation
+Substation is a a compute/data pipeline framework specifically designed for this project.
+It is designed to be useful enough to be applicable outside this project - it may eventually move
+into its own package
