+++ openknotscore/substation/runners/slurm.py
@@ -14,6 +14,7 @@ from ..db import DBReader
 
 @dataclass
 class SlurmRunner(Runner):
+    db_path: str
     partitions: str
     max_cores: int
     max_mem_per_core: int
@@ -71,7 +72,7 @@ class SlurmRunner(Runner):
             ) for alloc in config.allocations
         )
 
-    def run(self, tasks, job_name, config):
+    def run(self, tasks, job_name):
         schedule: Schedule = schedule_tasks(
             tasks,
             [
@@ -85,7 +86,7 @@ class SlurmRunner(Runner):
             ]
         )
         
-        dbpath = self.serialize_tasks(schedule, job_name, config)
+        dbpath = self.serialize_tasks(schedule, job_name, self.db_path)
         
         allocated_jobs = 0
         for idx, comp_config in enumerate(schedule.compute_configurations):
@@ -98,9 +99,9 @@ class SlurmRunner(Runner):
 
             max_gpus = self.config_max_gpus(comp_config)
             sbatch(
-                f'python -c "from openknotscore.runners.slurm import SlurmRunner; SlurmRunner.run_serialzied_allocation(\'{dbpath}\', {idx}, {'$SLURM_ARRAY_TASK_ID' if array_size > 1 else 0})"',
+                f'python -c "from openknotscore.substation.runners.slurm import SlurmRunner; SlurmRunner.run_serialzied_allocation(\'{dbpath}\', {idx}, {'$SLURM_ARRAY_TASK_ID' if array_size > 1 else 0})"',
                 f'{job_name}-{idx}' if len(schedule.compute_configurations) > 1 else job_name,
-                path.join(config.db_path, f'slurm-logs'),
+                path.join(self.db_path, f'slurm-logs'),
                 timeout=self.config_max_runtime(comp_config),
                 partition=self.partitions,
                 cpus=self.config_max_cpus(comp_config),
@@ -112,7 +113,7 @@ class SlurmRunner(Runner):
                 echo_cmd=True
             )
 
-    def forecast(self, tasks, config):        
+    def forecast(self, tasks):        
         schedule: Schedule = schedule_tasks(
             tasks,
             [
@@ -160,7 +161,7 @@ class SlurmRunner(Runner):
     def _srun_queue(dbpath: str, queue_id: int, queue: DBQueue, finished_queues: multiprocessing.Queue):
         print('srun queue', queue_id)
         srun(
-            ['python', '-c', f'from openknotscore.runners.runner import Runner; Runner.run_serialized_queue("{dbpath}", {queue_id})'],
+            ['python', '-c', f'from openknotscore.substation.runners.runner import Runner; Runner.run_serialized_queue("{dbpath}", {queue_id})'],
             cpus=queue.cpus,
             gpu_cmode='shared' if queue.gpu_id != '' else None,
             cuda_visible_devices=queue.gpu_id if queue.gpu_id != '' else None
@@ -205,7 +206,7 @@ def srun(
 
     args += command
 
-    if os.environ.get('OKSP_RUNNER_DRY_RUN') == 'true':
+    if os.environ.get('SUBSTATION_RUNNER_DRY_RUN') == 'true':
         print(f'DRY RUN SRUN COMMAND: {args}\n')
         return 0
     else:
@@ -278,7 +279,7 @@ def sbatch(
 
     input = f'{shebang}\n\n{set}\n\n{body}'
 
-    if os.environ.get('OKSP_RUNNER_DRY_RUN') == 'true':
+    if os.environ.get('SUBSTATION_RUNNER_DRY_RUN') == 'true':
         print(f'DRY RUN SBATCH STDIN:\n{input}\n-------')
         print(f'DRY RUN SBATCH COMMAND: {args}\n--------------')
     else:
